+++ src/molecules/EventCard/index.js
@@ -1,7 +1,22 @@
 import styles from './EventCard.module.css';
-import React from 'react';
+import React, { useCallback } from 'react';
 
-export default function EventCard({ date, type, message }) {
+/**
+ * @param {object} props
+ * @param {number} props.id
+ * @param {number} props.date unix timestamp in seconds
+ * @param {'ERROR' | 'NOTICE'} props.type
+ * @param {string} props.message
+ * @param {(id: number) => void} props.onArchive
+ * @returns
+ */
+export default function EventCard({ id, date, type, message, onArchive }) {
+	const onArchiveHandler = useCallback(() => {
+		if (typeof onArchive === 'function') {
+			onArchive(id);
+		}
+	}, [id, onArchive]);
+
 	return (
 		<div className={styles.container + ' ' + styles[type.toLowerCase()]}>
 			<div className={styles.header}>
@@ -11,6 +26,11 @@ export default function EventCard({ date, type, message }) {
 					hour: 'numeric',
 					minute: 'numeric'
 				})}</time>
+				{typeof onArchive === 'function' && (
+					<nav>
+						<button onClick={onArchiveHandler}>Archive</button>
+					</nav>
+				)}
 			</div>
 			<div className={styles.body}>
 				<p>{message}</p>