+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.35.0] - 2022-10-18
+### Added
+- Event archive button.
+
+### Changed
+- Updated npm packages.
+
+### Fixed
+- Typo label week electricity chart.
+- Refreshing some widgets on revisit page.
+
 ## [1.34.0] - 2022-09-30
 ### Added
 - Added solar to week chart.